requirements.txt
pyTelegramBotAPI==4.29.1
python-dotenv
requests
beautifulsoup4